Rename top-level module from datazilla to dzclient.
