+++ dzclient/__init__.py
@@ -2,4 +2,4 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from datazilla import DatazillaResult, DatazillaRequest
+from .client import DatazillaResult, DatazillaRequest
