+++ dzclient/tests/test_datazilla_result.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from ..datazilla import DatazillaResult
+from dzclient import DatazillaResult
