+++ setup.py
@@ -24,9 +24,9 @@ setup(name='datazilla',
       author_email='mdas@mozilla.com',
       url='https://github.com/mozilla/datazilla_client',
       license='MPL',
-      packages=['datazilla'],
+      packages=['dzclient'],
       zip_safe=False,
       install_requires=deps,
-      test_suite='datazilla.tests',
+      test_suite='dzclient.tests',
       tests_require=["mock"],
       )
